HW#4 Keep all application properties in one file
