+++ student-test/src/main/java/ru/otus/application/configuration/ApplicationProperties.java
@@ -42,4 +42,37 @@ public class ApplicationProperties {
 	public String getAnswerIdSeparator() {
 		return  separator.getAnswerId();
 	}
+
+	private static class CsvProperties {
+		private String path;
+
+		String getPath() {
+			return path;
+		}
+
+		public void setPath(String path) {
+			this.path = path;
+		}
+	}
+
+	private static class SeparatorProperties {
+		private String main;
+		private String answerId;
+
+		String getMain() {
+			return main;
+		}
+
+		public void setMain(String main) {
+			this.main = main;
+		}
+
+		String getAnswerId() {
+			return answerId;
+		}
+
+		public void setAnswerId(String answerId) {
+			this.answerId = answerId;
+		}
+	}
 }
