--- student-test/src/main/java/ru/otus/application/configuration/CsvProperties.java
@@ -1,13 +1,0 @@
-package ru.otus.application.configuration;
-
-public class CsvProperties {
-	private String path;
-
-	String getPath() {
-		return path;
-	}
-
-	public void setPath(String path) {
-		this.path = path;
-	}
-}
