--- student-test/src/main/java/ru/otus/application/configuration/SeparatorProperties.java
@@ -1,22 +1,0 @@
-package ru.otus.application.configuration;
-
-public class SeparatorProperties {
-	private String main;
-	private String answerId;
-
-	String getMain() {
-		return main;
-	}
-
-	public void setMain(String main) {
-		this.main = main;
-	}
-
-	String getAnswerId() {
-		return answerId;
-	}
-
-	public void setAnswerId(String answerId) {
-		this.answerId = answerId;
-	}
-}
